inventory-microservice/requirements.txt
anyio==3.7.0
async-timeout==4.0.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
hiredis==2.2.3
httptools==0.5.0
idna==3.4
more-itertools==8.14.0
pptree==3.1
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
python-ulid==1.1.0
PyYAML==6.0
redis==4.5.5
redis-om==0.1.2
sniffio==1.3.0
starlette==0.27.0
types-pyOpenSSL==23.2.0.0
types-redis==4.5.5.2
typing_extensions==4.6.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
